requirements.txt
Django==4.0.4
djangorestframework==3.14.0
pandas==1.3.2
sentence_transformers==2.2.2
transformers==4.22.1
